+++ crates/swc/src/builder.rs
@@ -237,7 +237,7 @@ impl<'a, 'b, P: swc_ecma_visit::Fold> PassBuilder<'a, 'b, P> {
                         nullish_coalescing: compat::es2020::nullish_coalescing::Config {
                             no_document_all: assumptions.no_document_all
                         },
-                        optional_chaining: compat::es2020::opt_chaining::Config {
+                        optional_chaining: compat::es2020::optional_chaining::Config {
                             no_document_all: assumptions.no_document_all,
                             pure_getter: assumptions.pure_getters
                         }
