+++ crates/swc_ecma_transforms_compat/src/es2020/optional_chaining.rs
