+++ crates/swc_ecma_transforms_compat/tests/es2020_optional_chaining.rs
@@ -1,7 +1,7 @@
 use std::{fs::read_to_string, path::PathBuf};
 
 use swc_ecma_parser::Syntax;
-use swc_ecma_transforms_compat::es2020::{opt_chaining::Config, optional_chaining};
+use swc_ecma_transforms_compat::es2020::{optional_chaining, optional_chaining::Config};
 use swc_ecma_transforms_testing::{compare_stdout, test, test_exec, test_fixture};
 use swc_ecma_visit::Fold;
